moves project file into project folder
